+++ src/Elasticsearch/Mapping/Drivers/Factories/Properties/Numeric/LongTypeFactory.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Elasticsearch\Mapping\Drivers\Factories\Properties\Numeric;
+
+use Elasticsearch\Mapping\Drivers\Factories\Properties\PropertyFactoryInterface;
+use Elasticsearch\Mapping\Types\AbstractType;
+use Elasticsearch\Mapping\Types\Common\Numeric\LongType;
+use stdClass;
+
+class LongTypeFactory implements PropertyFactoryInterface
+{
+    use NumericConfigurationTrait;
+
+    public static function create(string $name, stdClass $configuration): AbstractType
+    {
+        $longType = new LongType();
+        $longType->setName($name);
+        self::applyConfiguration($longType, $configuration);
+
+        return $longType;
+    }
+}