+++ src/Elasticsearch/Mapping/Drivers/Resolvers/PropertiesResolver/PropertiesResolver.php
@@ -4,10 +4,24 @@ declare(strict_types=1);
 
 namespace Elasticsearch\Mapping\Drivers\Resolvers\PropertiesResolver;
 
+use Elasticsearch\Mapping\Drivers\Factories\Properties\AliasTypeFactory;
+use Elasticsearch\Mapping\Drivers\Factories\Properties\BinaryTypeFactory;
 use Elasticsearch\Mapping\Drivers\Factories\Properties\BooleanTypeFactory;
+use Elasticsearch\Mapping\Drivers\Factories\Properties\Dates\DateNanoTypeFactory;
+use Elasticsearch\Mapping\Drivers\Factories\Properties\Dates\DateTypeFactory;
+use Elasticsearch\Mapping\Drivers\Factories\Properties\Keywords\ConstantKeywordTypeFactory;
 use Elasticsearch\Mapping\Drivers\Factories\Properties\Keywords\KeywordTypeFactory;
+use Elasticsearch\Mapping\Drivers\Factories\Properties\Keywords\WildcardTypeFactory;
+use Elasticsearch\Mapping\Drivers\Factories\Properties\Numeric\ByteTypeFactory;
+use Elasticsearch\Mapping\Drivers\Factories\Properties\Numeric\DoubleTypeFactory;
 use Elasticsearch\Mapping\Drivers\Factories\Properties\Numeric\FloatTypeFactory;
+use Elasticsearch\Mapping\Drivers\Factories\Properties\Numeric\HalfFloatTypeFactory;
 use Elasticsearch\Mapping\Drivers\Factories\Properties\Numeric\IntegerTypeFactory;
+use Elasticsearch\Mapping\Drivers\Factories\Properties\Numeric\LongTypeFactory;
+use Elasticsearch\Mapping\Drivers\Factories\Properties\Numeric\ScaledFloatTypeFactory;
+use Elasticsearch\Mapping\Drivers\Factories\Properties\Numeric\ShortTypeFactory;
+use Elasticsearch\Mapping\Drivers\Factories\Properties\Numeric\UnsignedLongTypeFactory;
+use Elasticsearch\Mapping\Drivers\Factories\Properties\Text\MatchOnlyTextTypeFactory;
 use Elasticsearch\Mapping\Drivers\Factories\Properties\Text\TextTypeFactory;
 use Elasticsearch\Mapping\Index;
 use Elasticsearch\Mapping\Types\Helpers\MultiFieldsInterface;
@@ -20,11 +34,25 @@ final class PropertiesResolver
 {
     /** @var class-string<PropertyFactoryInterface>[] */
     private array $propertiesFactories = [
-        'text'    => TextTypeFactory::class,
-        'keyword' => KeywordTypeFactory::class,
-        'integer' => IntegerTypeFactory::class,
-        'float'   => FloatTypeFactory::class,
-        'boolean' => BooleanTypeFactory::class,
+        'text'             => TextTypeFactory::class,
+        'match_only_text'  => MatchOnlyTextTypeFactory::class,
+        'keyword'          => KeywordTypeFactory::class,
+        'constant_keyword' => ConstantKeywordTypeFactory::class,
+        'wildcard'         => WildcardTypeFactory::class,
+        'integer'          => IntegerTypeFactory::class,
+        'long'             => LongTypeFactory::class,
+        'short'            => ShortTypeFactory::class,
+        'byte'             => ByteTypeFactory::class,
+        'unsigned_long'    => UnsignedLongTypeFactory::class,
+        'float'            => FloatTypeFactory::class,
+        'double'           => DoubleTypeFactory::class,
+        'half_float'       => HalfFloatTypeFactory::class,
+        'scaled_float'     => ScaledFloatTypeFactory::class,
+        'boolean'          => BooleanTypeFactory::class,
+        'date'             => DateTypeFactory::class,
+        'date_nanos'       => DateNanoTypeFactory::class,
+        'binary'           => BinaryTypeFactory::class,
+        'alias'            => AliasTypeFactory::class,
     ];
 
     /**